feat: S3 코드 추가를 위한 기존 이미지 로직 추상화, 구현체 변경
